[NCL-1983] Allow SCMImpl to clone different repositories in parallel

+++ source-code-manager/src/main/java/org/jboss/da/scm/impl/SCMImpl.java
@@ -1,27 +1,24 @@
 package org.jboss.da.scm.impl;
 
-import org.jboss.da.scm.api.SCMType;
+import org.apache.maven.scm.ScmException;
 import org.jboss.da.scm.api.SCM;
+import org.jboss.da.scm.api.SCMType;
+
+import javax.inject.Inject;
+
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
-import java.util.HashMap;
-import java.util.Iterator;
-import java.util.Map;
-import java.util.Optional;
-import javax.annotation.PreDestroy;
-import javax.ejb.Schedule;
-import javax.inject.Inject;
-import javax.inject.Singleton;
-import org.apache.commons.io.FileUtils;
-import org.apache.maven.scm.ScmException;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
 import org.slf4j.Logger;
+import org.apache.commons.io.FileUtils;
 
 /**
  *
  * @author Honza Brázdil <jbrazdil@redhat.com>
  */
-@Singleton
 public class SCMImpl implements SCM {
 
     @Inject
@@ -30,59 +27,46 @@ public class SCMImpl implements SCM {
     @Inject
     ScmFacade scm;
 
-    private final Map<SCMSpecifier, DirectoryReference> cache = new HashMap<>();
+    @Inject
+    SCMCache cache;
 
     @Override
     public File cloneRepository(SCMType scmType, String scmUrl, String revision)
             throws ScmException {
-        try {
-            SCMSpecifier spec = new SCMSpecifier(scmType, scmUrl, revision);
+        SCMSpecifier spec = new SCMSpecifier(scmType, scmUrl, revision);
 
-            if (cache.containsKey(spec)) {
-                Optional<File> file = cache.get(spec).get();
-                if (file.isPresent()) {
-                    log.debug("Found cached repository for {} at {}", spec, file);
-                    return file.get();
-                }
-            }
+        FutureReference fref = cache.get(spec);
 
-            File tempDir = Files.createTempDirectory("cloned_repo").toFile();
-            log.debug("Cached repository for {} not found. Cloning to {}", spec, tempDir);
+        if (fref.shouldIComplete()) {
             try {
-                scm.shallowCloneRepository(scmType, scmUrl, revision, tempDir);
-                DirectoryReference ref = new DirectoryReference(tempDir);
-                cache.put(spec, ref);
-                return tempDir;
-            } catch (ScmException ex) {
+                File tempDir = Files.createTempDirectory("cloned_repo").toFile();
+                log.debug("Cached repository for {} not found. Cloning to {}", spec, tempDir);
                 try {
-                    FileUtils.deleteDirectory(tempDir);
-                } catch (IOException ioex) {
-                    log.warn("Temporary directory could not be deleted", ex);
+                    scm.shallowCloneRepository(scmType, scmUrl, revision, tempDir);
+                    DirectoryReference ref = new DirectoryReference(tempDir);
+                    fref.complete(ref);
+                    return tempDir;
+                } catch (ScmException ex) {
+                    fref.completeExceptionally(ex);
+                    try {
+                        FileUtils.deleteDirectory(tempDir);
+                    } catch (IOException ioex) {
+                        
+                        log.warn("Temporary directory could not be deleted", ioex);
+                    }
+                    throw ex;
                 }
-                throw ex;
+            } catch (IOException ex) {
+                throw new ScmException(
+                        "Could not create temp directory for cloning the repository", ex);
             }
-        } catch (IOException ex) {
-            throw new ScmException("Could not create temp directory for cloning the repository", ex);
-        }
-    }
 
-    @Schedule(hour = "*", minute = "*/10")
-    public void invalidateCache() {
-        Iterator<Map.Entry<SCMSpecifier, DirectoryReference>> it = cache.entrySet().iterator();
-        while (it.hasNext()) {
-            DirectoryReference ref = it.next().getValue();
-            if (!ref.check()) {
-                it.remove();
+        } else {
+            try {
+                return fref.get(30, TimeUnit.MINUTES).get().orElseThrow(()->new IllegalStateException("Now completed reference has empty file."));
+            } catch (InterruptedException | ExecutionException | TimeoutException ex) {
+                throw new ScmException("Could not obtain cloned repository.", ex);
             }
-        }
-    }
-
-    @PreDestroy
-    private void cleanup() {
-        Iterator<Map.Entry<SCMSpecifier, DirectoryReference>> it = cache.entrySet().iterator();
-        while (it.hasNext()) {
-            it.next().getValue().delete();
-            it.remove();
         }
     }
 }
